Refactor handling of string arrays
Refactor code to handle passing string arrays for gas names from F90 to
C++. Instead of needing a separate interface to add gases one by one to
the C++ interface, pass a C pointer to an array of null-terminated
strings to the C++ interface. To simplify this, we add a function that
takes an F90 string array, appends the C_NULL_CHAR to the end of each
element, and returns a C pointer. To workaround a strange issue with
moving this to a function, we also need to return the array of converted
strings that the pointer references. All this allows us to nearly
directly call the C++ version of the rrtmgpxx_initialize and run
routines, and remove quite a bit of boiler plate code.